LF-5390 Relocate shared MenuItem component out of legacy OrganicCertifierSurvey tree
